// src/shared/api/generated/model/authSession.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vendora API
 * Vendora v1 marketplace API optimized for OTP-first authentication and frontend-driven commerce.
 * OpenAPI spec version: 1.0.0
 */
import type { User } from './user';
import type { AuthTokens } from './authTokens';

export interface AuthSession {
  user: User;
  tokens: AuthTokens;
  onboardingCompleted: boolean;
}