// src/shared/api/generated/model/authTokens.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vendora API
 * Vendora v1 marketplace API optimized for OTP-first authentication and frontend-driven commerce.
 * OpenAPI spec version: 1.0.0
 */

export interface AuthTokens {
  accessToken: string;
  refreshToken: string;
  /** @minimum 1 */
  expiresIn: number;
}